// map-app/src/types/game.ts
export interface GameMapMeta {
  id: string;
  name: string;
  type: string;
  tileWidth: number;
  tileHeight: number;
  tilesCountX: number;
  tilesCountY: number;
  isVisible: boolean;
}
// ...
export type UserMarkerLocalType =
  | "completed"
  | "location"
  | "fox"
  | "favorite"
  | "gathering"
  | "creature";

export interface UserMarkerInstance {
  id: string;
  markerId: string;
  subtype: string;
  mapId: string;
  x: number;
  y: number;
  name: string;
  description: string;
  image: string;
  type: "local" | "feedback" | "uploaded";
  localType?: UserMarkerLocalType;
  status?: "pending" | "accepted" | "rejected" | "revision" | "deleted";
  reply?: string;
}